requirements.txt
absl-py==2.1.0
action-msgs==2.0.3
action-tutorials-interfaces==0.33.6
action-tutorials-py==0.33.6
actionlib-msgs==5.3.6
actuator-msgs==0.0.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
ament-cmake-test==2.5.4
ament-copyright==0.17.3
ament-cppcheck==0.17.3
ament-cpplint==0.17.3
ament-flake8==0.17.3
ament-index-python==1.8.1
ament-lint==0.17.3
ament-lint-cmake==0.17.3
ament-package==0.16.4
ament-pep257==0.17.3
ament-uncrustify==0.17.3
ament-xmllint==0.17.3
angles==1.16.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==24.2.0
audioread==3.0.1
Brotli @ file:///croot/brotli-split_1714483155106/work
builtin-interfaces==2.0.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1752493286596/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work

click==8.2.1
colorlog==6.9.0
composition-interfaces==2.0.3
contourpy==1.3.0
control-msgs==5.5.0
controller-manager==4.35.0
controller-manager-msgs==4.35.0
cv-bridge==4.1.0
cycler==0.12.1
decorator==5.2.1
demo-nodes-py==0.33.6
diagnostic-msgs==5.3.6
diagnostic-updater==4.2.6
docopt==0.6.2
domain-coordinator==0.12.0
einops==0.8.1
example-interfaces==0.12.0
examples-rclpy-executors==0.19.6
examples-rclpy-minimal-action-client==0.19.6
examples-rclpy-minimal-action-server==0.19.6
examples-rclpy-minimal-client==0.19.6
examples-rclpy-minimal-publisher==0.19.6
examples-rclpy-minimal-service==0.19.6
examples-rclpy-minimal-subscriber==0.19.6
ffmpeg-image-transport-msgs==1.0.2
filelock @ file:///croot/filelock_1700591183607/work
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.7.0
fsspec==2025.9.0
generate-parameter-library-py==0.5.0
geographic-msgs==1.0.6
geometry-msgs==5.3.6
gevent==25.5.1
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
gps-msgs==2.1.1
greenlet==3.2.3
hf-xet==1.1.9
huggingface-hub==0.34.4
HyperPyYAML==1.2.2
idna @ file:///croot/idna_1714398848350/work
image-geometry==4.1.0
interactive-markers==2.5.4
jax==0.4.35
jaxlib==0.4.35
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.5.2
julius==0.2.7
keyboard==0.13.5
kiwisolver==1.4.7
laser-geometry==2.7.1
launch==3.4.6
launch-ros==0.26.8
launch-testing==3.4.6
launch-testing-ros==0.26.8
launch-xml==3.4.6
launch-yaml==3.4.6
lazy_loader==0.4
librosa==0.11.0
lifecycle-msgs==2.0.3
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.44.0
logging-demo==0.33.6
Mako==1.3.10
map-msgs==2.4.1
markdown-it-py==4.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.15
message-filters==4.11.7
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
ml_dtypes==0.5.0
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.1
multidict==6.6.4
nav-msgs==5.3.6
networkx @ file:///croot/networkx_1720002482208/work
nmea-msgs==2.1.0
numba==0.61.2
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optuna==4.5.0
osrf-pycommon==2.1.7
packaging==24.1
pal-statistics==2.6.4
pal-statistics-msgs==2.6.4
pandas==2.2.3
pcl-msgs==1.0.0
pendulum-msgs==0.33.6
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.4.0
point-cloud-interfaces==4.0.1
pooch==1.8.2
primePy==1.3
propcache==0.3.2
protobuf==4.25.5
pvporcupine==3.0.5
pvrecorder==1.2.7
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
PyAudio==0.2.14
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-qt-binding==2.2.1
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.6
rcl-interfaces==2.0.3
rclpy==7.1.5
rcutils==6.7.3
regex==2025.9.1
requests @ file:///croot/requests_1721410876868/work
resource-retriever==3.4.4
rich==14.1.0
rmw-dds-common==3.1.0
robot-localization==3.8.3
ros-gz-bridge==1.0.16
ros-gz-interfaces==1.0.16
ros-gz-sim==1.0.16
ros2action==0.32.5
ros2bag==0.26.9
ros2bag-mcap-cli==0.26.9
ros2bag-sqlite3-cli==0.26.9
ros2cli==0.32.5
ros2component==0.32.5
ros2controlcli==4.35.0
ros2doctor==0.32.5
ros2interface==0.32.5
ros2launch==0.26.8
ros2lifecycle==0.32.5
ros2multicast==0.32.5
ros2node==0.32.5
ros2param==0.32.5
ros2pkg==0.32.5
ros2run==0.32.5
ros2service==0.32.5
ros2topic==0.32.5
rosbag2-interfaces==0.26.9
rosbag2-py==0.26.9
rosgraph-msgs==2.0.3
rosidl-adapter==4.6.6
rosidl-cli==4.6.6
rosidl-cmake==4.6.6
rosidl-generator-c==4.6.6
rosidl-generator-cpp==4.6.6
rosidl-generator-py==0.22.1
rosidl-generator-type-description==4.6.6
rosidl-parser==4.6.6
rosidl-pycommon==4.6.6
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.2
rosidl-typesupport-fastrtps-cpp==3.6.2
rosidl-typesupport-introspection-c==4.6.6
rosidl-typesupport-introspection-cpp==4.6.6
rpyutils==0.4.1
rqt-action==2.2.0
rqt-bag==1.5.5
rqt-bag-plugins==1.5.5
rqt-console==2.2.1
rqt-graph==1.5.5
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-msg==1.5.1
rqt-plot==1.4.4
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.3
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-learn==1.7.1
scipy @ file:///croot/scipy_1717521478074/work/dist/scipy-1.13.1-cp310-cp310-linux_x86_64.whl#sha256=c73757af21f88127db3e72c16ee95d5a104ebd533b662667d352c0765772cf83
semver==3.0.4
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
sentencepiece==0.2.1
service-msgs==2.0.3
shape-msgs==5.3.6
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==2.0.43
sros2==0.13.4
statistics-msgs==2.0.3
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
sympy==1.13.1
tabulate==0.9.0
teleop-twist-keyboard==2.4.0
tensorboardX==2.6.4
tf-transformations==1.1.0
tf2-geometry-msgs==0.36.14
tf2-kdl==0.36.14
tf2-msgs==0.36.14
tf2-py==0.36.14
tf2-ros-py==0.36.14
tf2-sensor-msgs==0.36.14
tf2-tools==0.36.14
theora-image-transport==4.0.5
threadpoolctl==3.6.0
tokenizers==0.22.0
tomli==2.2.1
topic-monitor==0.33.6
torch==2.5.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.8.2
torchvision==0.20.1
tqdm==4.66.6
trajectory-msgs==5.3.6
transformers==4.56.0
triton==3.1.0
turtlebot3-msgs==2.4.0
turtlesim==1.8.3
type-description-interfaces==2.0.3
typer==0.17.4
typing_extensions==4.15.0
tzdata==2024.2
unique-identifier-msgs==2.5.0
urllib3 @ file:///croot/urllib3_1727769808118/work
vision-msgs==4.1.1
visualization-msgs==5.3.6
webrtcvad==2.0.10
websocket-client==1.8.0
xacro==2.0.13
yarl==1.20.1
zed-msgs==5.0.1
zope.event==5.1.1
zope.interface==7.2
